requirements.txt
beautifulsoup4
celery==5.4.0
dataclasses==0.6
dataclasses-json==0.6.7
fastapi==0.113.0
loguru
pandas
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
sqlalchemy
ulid
uvicorn==0.30.6
uvloop==0.19.0
websockets==12.0
